requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
# FastAPI 0.104 / Starlette 0.27 的 TestClient 与 httpx 0.28+ 不兼容。
# 锁到 0.27.x，避免本地新环境里测试全部失效。
httpx>=0.27,<0.28
pydantic>=2.8.0
python-dotenv==1.0.0

# LLM and Embeddings
openai>=1.12.0
anthropic>=0.23.0
transformers>=4.57.0,<5.0.0
torch>=2.1.1
peft>=0.7.1
sentence-transformers>=5.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langgraph>=0.2.0

# Vector DB and RAG (llama-index 0.10+ with required integrations)
llama-index>=0.10.58
llama-index-vector-stores-qdrant>=0.2.0
llama-index-retrievers-bm25>=0.3.0
llama-index-postprocessor-flag-embedding-reranker>=0.1.3
qdrant-client>=1.7.0

# Database
supabase>=2.4.0
# 2.9.9 在 Python 3.13 Windows 环境下容易回退到源码构建并安装失败。
# 放宽到支持新版本解释器 wheel 的范围，降低本地开发环境阻塞概率。
psycopg2-binary>=2.9.10,<3.0.0

# Data Processing
pandas>=2.1.3
numpy>=1.24.3
scikit-learn>=1.3.2
rouge-score>=0.1.2

# Utilities
requests>=2.31.0
schedule>=1.2.0
aiohttp>=3.9.1
python-multipart>=0.0.6
boto3>=1.34.0
jinja2>=3.1.4
plotly>=5.24.0

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.12.0
pylint>=3.0.3

# Logging
python-json-logger>=2.0.7
